% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_categorical_interaction_mappings.R
\name{apply.categorical.interaction.mappings}
\alias{apply.categorical.interaction.mappings}
\title{Applies categorical interacion mappins}
\usage{
apply.categorical.interaction.mappings(data,
  categorical.interactions.mappings, verbose = TRUE)
}
\arguments{
\item{data}{[required | data.frame] Dataset containing categorical features}

\item{categorical.interactions.mappings}{[ required | list] Mapping list produced by function map.categorical.interactions.}

\item{verbose}{[optional | logical | default=TRUE] Chatty or silent function output}
}
\value{
A data.frame object with interacted features
}
\description{
Creates categorical interaction features by combining categorical features.
}
\examples{
cim <- map.categorical.interactions(x = c("Class","Sex","Age"))
cints <- apply.categorical.interaction.mappings(data=as.data.frame(Titanic), categorical.interactions.mappings = cim)
}
\author{
Xander Horn
}
