% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_impute_mappings.R
\name{apply.impute.mappings}
\alias{apply.impute.mappings}
\title{Applies imputation mappings}
\usage{
apply.impute.mappings(data, impute.mappings, impute.mode = "auto",
  na.threshold = 0.1, track.features = TRUE, progress = TRUE)
}
\arguments{
\item{data}{[required | data.frame] Dataset containing features to impute}

\item{impute.mappings}{[required | list] Output from function map.impute}

\item{impute.mode}{[optional | character | default="auto"] Imputation mode, options are auto, encode and median.mode. Auto applies a combination bewteen encoding and median.mode imputation based on the na.threshold parameter}

\item{na.threshold}{[optional | numeric | default=0.1] Threshold for auto impute.mode to apply encoding or median.mode imputation. All features containing Na values above the specified percentage threshold will be imputed using encoding}

\item{track.features}{[optional | logical | default=TRUE] Creates tracking features that records which observations had NA values present}

\item{progress}{[optional | logical | default=TRUE] Display a progress bar}
}
\value{
Data frame with imputed original features and tracking features
}
\description{
Uses output from map.impute to impute NA values in a dataframe. Also creates tracking features to indicate which observations had NA values.
}
\examples{
imp <- map.impute(data = iris)
new_iris <- apply.impute.mappings(iris, imp)
}
\author{
Xander Horn
}
