% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.kmeans.mappings.R
\name{apply.kmeans.mappings}
\alias{apply.kmeans.mappings}
\title{Applies kmeans feature mappings}
\usage{
apply.kmeans.mappings(data, kmeans.mappings, progress = TRUE)
}
\arguments{
\item{data}{[required | data.frame] Dataset containing features to apply mappings to}

\item{kmeans.mappings}{[required | list] Output from function map.kmeans.features}

\item{progress}{[optional | logical | default=TRUE] Display a progress bar}
}
\value{
Data frame with newly added features and original features
}
\description{
Applies mapped kmeans features by calculating the distance to center per observation
}
\examples{
km <- map.kmeans.features(data = iris, x = setdiff(names(iris), "Species"))
new_iris <- apply.kmeans.mappings(data = iris, kmeans.mappings = km)
}
\author{
Xander Horn
}
