% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_numeric_interactions.R
\name{apply.numeric.interactions}
\alias{apply.numeric.interactions}
\title{Creates numeric interaction features}
\usage{
apply.numeric.interactions(data, numeric.combination.frame,
  verbose = TRUE)
}
\arguments{
\item{data}{[required | data.frame] A data.frame object on which the interactions should be applied to containing the base features used to create the combination of interaction features.}

\item{numeric.combination.frame}{[required | data.frame] A data.frame object containing a combination of features used to created interactions from.}

\item{verbose}{[optional | logical] Chatty or silent function output}
}
\value{
A data.frame object containing only interacted features
}
\description{
Creates interaction features off of numeric features by summing, dividing, multiplying and deducting. Interactions are created based off of a data.frame of combinations from function map.numeric.features.
}
\examples{
nim <- map.numeric.interactions(x=names(iris)[1:4])
nints <- apply.numeric.interactions(data = iris, numeric.combination.frame = nim)
}
\author{
Xander Horn
}
