% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_outlier_mappings.R
\name{apply.outlier.mappings}
\alias{apply.outlier.mappings}
\title{Applies outlier mapping tables}
\usage{
apply.outlier.mappings(data, outlier.mappings, progress = TRUE,
  track.features = TRUE)
}
\arguments{
\item{data}{[required | data.frame] Dataset containing features}

\item{outlier.mappings}{[required | list] Output from function map.outliers}

\item{progress}{[optional | logical | default=TRUE] Display a progress bar}

\item{track.features}{[optional | logical | default=TRUE] Creates tracking features that records which observations had outliers present}
}
\value{
Data frame with clipped features and all other features
}
\description{
Clips outliers by applying mapping tables to respective features. Also creates tracking features to identify which observations had outliers present for specific features.
}
\examples{
om <- map.outliers(iris, x = names(iris)[1:4])
new_iris <- apply.outlier.mappings(data, om)
}
\author{
Xander Horn
}
