% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automl.R
\name{automl}
\alias{automl}
\title{Automated machine learning}
\usage{
automl(train, y, valid = NULL, test = NULL, x = NULL,
  id.feats = NULL, time.partition.feature = NULL,
  optimization.metric = "AUTO", valid.split = 0.1, test.split = 0.2,
  pipeline.search.max.runtime.mins = 30,
  automl.search.max.runtime.mins = 30, balance.classes = FALSE,
  models = c("DRF", "GLM", "GBM", "XGBoost", "DeepLearning",
  "StackedEnsemble"), cv.folds = 0, max.levels = 100,
  data.leakage.cutoff = 0.65, cluster.memory = NULL,
  min.feature.importance = 0.1, seed = 1, output.path = NULL,
  pipeline = NULL, return.data = FALSE)
}
\arguments{
\item{train}{[required | data.frame] Traning set, if no test and validation sets are provided it is considered as the full set and test and validation sets will be created.}

\item{y}{[optional | character] The name of the target feature contained in the training and validation sets.}

\item{valid}{}

\item{test}{}

\item{x}{}

\item{id.feats}{}

\item{time.partition.feature}{}

\item{optimization.metric}{}

\item{valid.split}{}

\item{test.split}{}

\item{pipeline.search.max.runtime.mins}{}

\item{automl.search.max.runtime.mins}{}

\item{balance.classes}{}

\item{models}{}

\item{cv.folds}{}

\item{max.levels}{}

\item{data.leakage.cutoff}{}

\item{cluster.memory}{}

\item{min.feature.importance}{}

\item{seed}{}

\item{output.path}{}
}
\description{
Automated machine learning with automated feature engineering via pipeline exploration optimization. Utilises h2o.automl as the modelling engine. Duplicate observations are removed based on id features provided to the function. Time sensitive partitioning is also performed if a time sensitive indicator feature is provided. The function is bound by time for both optimization of pipelines as well as model optimization.
}
