% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_interactions.R
\name{categorical.interactions}
\alias{categorical.interactions}
\title{Categorical feature interactions}
\usage{
categorical.interactions(data, x, n.interactions = 2, progress = TRUE)
}
\arguments{
\item{data}{[required | data.frame] Dataset containing categorical features}

\item{x}{[required | character] A vector of categorical feature names present in the dataset}

\item{n.interactions}{[optional | numeric | Default=2] Number of features to interact, needs to be less than or equal to the number of features provided in x}

\item{progress}{[optional | logical | default=TRUE] Display a progress bar}
}
\value{
Data frame or vector of interacted features
}
\description{
Computes categorical feature interactions by joining categories together.
}
\examples{
res <- categorical.interactions(data = as.data.frame(Titanic), x = c('Class','Sex','Age'))
}
\author{
Xander Horn
}
