% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_leak.R
\name{data.leak}
\alias{data.leak}
\title{Data leakage detection}
\usage{
data.leak(train, test, id.feats = NULL, sample.size = 0.3,
  seed = 1234, progress = TRUE)
}
\arguments{
\item{train}{[required | data.frame] Training data}

\item{test}{[required | data.frame] Testing data}

\item{id.feats}{[optional | character | default=NULL] Names of ID features}

\item{sample.size}{[optional | numeric | default=0.3] Percentage to down sample data for decreased computation time}

\item{seed}{[optional | integer | default=1234] Random number seed for reproducable results}

\item{progress}{[Optional | logical | default=TRUE] Display a progress bar}
}
\value{
Data frame containing AUC per feature indicating data leakage
}
\description{
Fits a decision tree model to determine which features have data leakage
}
\examples{
train <- iris[1:65,]
test <- iris[66:nrow(iris),]
res <- data.leak(train = train, test = test)
}
\author{
Xander Horn
}
