% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_features.R
\name{date.features}
\alias{date.features}
\title{Engineers date and time features}
\usage{
date.features(data, x, progress = TRUE)
}
\arguments{
\item{data}{[required | data.frame] Dataset containing date features}

\item{x}{[required | character] A vector of date feature names present in the dataset}
}
\value{
Data frame with date and time engineered features
}
\description{
Engineers date and time features from date time features
}
\examples{
iris$date <- as.Date(now())
dt <- date.features(data = iris, x = "date")
}
\author{
Xander Horn
}
