% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_duplicate.R
\name{de.duplicate}
\alias{de.duplicate}
\title{De-duplicate dataframe}
\usage{
de.duplicate(data, id.feats)
}
\arguments{
\item{data}{[required | data.frame]}

\item{id.feats}{[required | character] Names of ID features or other features to de-duplicate by}
}
\value{
Data frame containing unique observations
}
\description{
Ensures only unique observations are kept according to specific features
}
\examples{
res <- de.duplicate(iris, id.feats = "Sepal.Length")
}
\author{
Xander Horn
}
