% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe}
\alias{describe}
\title{Tabular exploratory data analysis}
\usage{
describe(data, sample.size = 0.3, seed = 1, progress = TRUE)
}
\arguments{
\item{data}{[required | data.frame] Dataset containing categorical features}

\item{sample.size}{[optional | numeric | default=0.3] Percentage to down sample data for decreased computation time}

\item{seed}{[optional | integer | default=1] Random number seed for reproducable results}

\item{progress}{[optional | logical | default=TRUE] Display a progress bar}
}
\value{
Data frame containing statistics of features
}
\description{
Provides a high level overview of a dataset by providing some summary statistics.
}
\examples{
res <- describe(iris)
}
\author{
Xander Horn
}
