% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_features.R
\name{detect.feats}
\alias{detect.feats}
\title{Detect feature types}
\usage{
detect.feats(data, x = NULL, text.threshold = 100, sample.size = 0.3,
  progress = TRUE)
}
\arguments{
\item{data}{[required | data.frame] Dataset containing categorical features}

\item{x}{[required | character] A vector of numeric feature names present in the dataset}

\item{text.threshold}{[optional | integer | default=100] Maximum characters in a feature before it is identified as a text feature}

\item{sample.size}{[optional | numeric | default=0.3] Percentage to down sample data for decreased computation time}

\item{progress}{[optional | logical | default=TRUE] Display a progress bar}
}
\value{
Data.frame with categorical, numeric, date, and text feature types
}
\description{
Detects the type of features present in a data.frame
}
\examples{
res <- detect.feats(iris)
}
\author{
Xander Horn
}
