% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document.R
\name{document}
\alias{document}
\title{Dynamic documentation for pipeline settings}
\usage{
document(pipeline)
}
\arguments{
\item{pipeline}{[requied | list] A pipeline list object either returned from automl or design.pipeline.}
}
\value{
List containing documentation
}
\description{
Creates dynamic documentation for pipelines.
}
\examples{
p <- design.pipeline()
doc <- document(p)
}
\author{
Xander Horn
}
