% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda.R
\name{eda}
\alias{eda}
\title{Automated exploratory data analysis}
\usage{
eda(data, x = NULL, y = NULL, sample.size = 0.5, theme = 1,
  numeric.plot = "histogram", categorical.plot = "bar",
  pipeline = NULL)
}
\arguments{
\item{data}{[required | data.frame ] Dataset to visualize.}

\item{x}{[optional | character | default=NULL] Features to visualize specified as a character vector. If NULL then all features in the dataset will be used except for the target feature.}

\item{y}{[optional | character | default=NULL] Target feature to visualize. If NULL a univariate visialization will take place.}

\item{sample.size}{[optional | numeric | default=0.5] Sample size to down sample the data for faster exploration.}

\item{theme}{[optional | numeric | default=1] Color theme applied to plot, options range from 1 to 4.}

\item{numeric.plot}{[optional | character | default="histogram"] The type of plot to be produced. For numeric feature types histogram, density, boxplot and violin are available.}

\item{categorical.plot}{[optional | character | default="bar"] The type of plot to be produced. For categorical bar and stackedbar are available.}

\item{pipeline}{[optional | list | default=NULL] Pipeline used to pre-process the data for visualization. If NULL then a exploratory pipeline will be produced to pre-process the data.}
}
\value{
List containing plots, tabular summary exploration and the pipeline used to data pre-processing.
}
\description{
Explores a provided data set and returns a list of plots per feature as well as a summary table.
}
\examples{
res <- eda(data = iris)
res <- eda(data = iris, y = "Species")
}
\author{
Xander Horn
}
