% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_plot.R
\name{lazy.plot}
\alias{lazy.plot}
\title{Lazy ggplot plotting}
\usage{
lazy.plot(data, x = NULL, y = NULL, type = "histogram",
  transparency = 1, theme = 1)
}
\arguments{
\item{data}{[required | data.frame] Dataset containing predictor and / or target features.}

\item{x}{[optional | character | default=NULL] A vector of feature names present in the dataset used to predict the target feature. If NULL then all columns in the dataset is used.}

\item{y}{[required | character | default=NULL] The name of the target feature contained in the dataset.}

\item{type}{[optional | character | default="histogram"] The type of plot to be produced. For numeric feature types histogram, density, boxplot and violin are available. For categorical bar and stackedbar are available.}

\item{transparency}{[optional | numeric | default = 1] Transparency applied to plots.}

\item{theme}{[optional | numeric | default=1] Color theme applied to plot, options range from 1 to 4.}
}
\value{
Plot of ggplot2 type
}
\description{
Quickly produces specifc plots using the ggplot library. This function is exported but its main purpose is to be used in the eda function.
}
\examples{
lazy.plot(iris, x = "Sepal.Length", y = "Species", type = "density")
lazy.plot(iris, x = "Sepal.Length", y = "Species", type = "violin")
}
\author{
Xander Horn
}
