% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazy_predict.R
\name{lazy.predict}
\alias{lazy.predict}
\title{Preprocess and predict wrapper}
\usage{
lazy.predict(data, model.object, pipeline, mapping.list)
}
\arguments{
\item{data}{[required | data.frame] New data to produce predictions for.}

\item{model.object}{[required | h2o] The h2o model object.}

\item{pipeline}{[required | list | default=NULL] List object returned either by automl or preprocess, if left as NULL, it is assumed that a prediction is required on an already pre-processed dataset.}

\item{mapping.list}{[required | list | default=NULL] List object returned either by automl or preprocess, if left as NULL, it is assumed that a prediction is required on an already pre-processed dataset..}
}
\value{
A data.frame object with predictions.
}
\description{
Pre-processes data and predicts. Function to be used in scoring environment. For classification problems the result will contain both the predicted class and probabilities.
}
\examples{
## NOT RUN
res <- automl(train=iris, valid=iris, y="Species")
preds <- lazy.predict(data=iris, model.object=res$models[1], pipeline=res$pipeline, mapping.list=res$mapping.list)
}
\author{
Xander Horn
}
