% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_categorical_mappings.R
\name{map.categorical.interactions}
\alias{map.categorical.interactions}
\title{Maps categorical interaction combination frames}
\usage{
map.categorical.interactions(x, n.interactions = 2)
}
\arguments{
\item{x}{[required | character] A vector of categorical feature names present in the dataset}

\item{n.interactions}{[optional | numeric | Default=2] Number of features to interact, needs to be less than or equal to the number of features provided in x}
}
\value{
List of data.frames containing a combination of features to interact
}
\description{
Creates a list of data.frame for each level of interactions containing a combination of categorical features to interact.
}
\examples{
cim <- map.categorical.interactions(x = c("Class","Sex","Age"))
}
\author{
Xander Horn
}
