% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_freq_encoding.R
\name{map.freq.encoding}
\alias{map.freq.encoding}
\title{Frequency encoding}
\usage{
map.freq.encoding(data, x = NULL, progress = TRUE)
}
\arguments{
\item{data}{[required | data.frame] Dataset containing categorical features}

\item{x}{[required | character] A vector of categorical or numerical feature names present in the dataset}

\item{progress}{[optional | logical | default=TRUE] Display a progress bar}
}
\value{
List of data frames containing engineered mapping features
}
\description{
Creates a list of frequency encoded data.frames for each feature provided.
}
\examples{
res <- map.freq.encoding(iris, names(iris))
}
\author{
Xander Horn
}
