% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_impute.R
\name{map.impute}
\alias{map.impute}
\title{Imputation mappings}
\usage{
map.impute(data, x = NULL, progress = TRUE)
}
\arguments{
\item{data}{[required | data.frame] Dataset containing features}

\item{x}{[required | character] A vector of feature names present in the dataset}

\item{progress}{[optional | logical | default=TRUE] Display a progress bar}
}
\value{
List of data frames containing imputation mapping values
}
\description{
Creates imputation mapping values for all features provided
}
\examples{
res <- map.impute(data = iris, x = names(iris))
}
\author{
Xander Horn
}
