% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_kmeans_features.R
\name{map.kmeans.features}
\alias{map.kmeans.features}
\title{Kmeans distance to center feature mappings}
\usage{
map.kmeans.features(data, x, clusters = 3, sample.size = 0.3,
  seed = 1, progress = TRUE)
}
\arguments{
\item{data}{[required | data.frame] Dataset containing categorical features}

\item{x}{[required | character] A vector of categorical feature names present in the dataset}

\item{clusters}{[optional | integer | default=3] The number of clusters to create in each feature}

\item{sample.size}{[optional | numeric | default=0.3] Percentage to down sample data for decreased computation time}

\item{seed}{[optional | integer| default=1] The random number seed for reproducable results}

\item{progress}{[optional | logical | default=TRUE] Display a progress bar}
}
\value{
List of data frames containing mapping tables
}
\description{
Creates mapping tables for each numerical feature containing the center for each feature's min and max value. These tables can then be applied using the function 'apply.kmeans.mappings' to calculate the distance to cluster center for each feature. Each feature is scaled by converting it to a range between 0 and 1 before clustering.
}
\examples{
res <- map.kmeans.features(data = iris, x = setdiff(names(iris), "Species"))
}
\author{
Xander Horn
}
