% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_max_scaler.R
\name{map.max.scaler}
\alias{map.max.scaler}
\title{Maximum value scaler}
\usage{
map.max.scaler(data, x, progress = TRUE)
}
\arguments{
\item{data}{[required | data.frame] Dataset containing numeric features}

\item{x}{[required | character] A vector of numeric feature names present in the dataset}

\item{progress}{[optional | logical | default=TRUE] Display a progress bar}
}
\value{
Data frame with features and max value
}
\description{
Scales all numerical feautes to be between 0 and 1.
}
\examples{
res <- map.max.scaler(data = iris, x = names(iris)[1:4])
}
\author{
Xander Horn
}
