% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_numeric_interactions.R
\name{map.numeric.interactions}
\alias{map.numeric.interactions}
\title{Numeric feature interactions}
\usage{
map.numeric.interactions(x)
}
\arguments{
\item{x}{[required | character] A vector of numeric feature names present in the dataset}
}
\value{
Data frame with combinations of all features to be interacted
}
\description{
Creates a data frame containing two-way combinations of all features to be used when creating interactions
}
\examples{
nim <- map.numeric.interactions(x = names(iris)[1:4])
}
\author{
Xander Horn
}
