% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_outliers.R
\name{map.outliers}
\alias{map.outliers}
\title{Creates mapping tables for detecting outlier values in numeric features}
\usage{
map.outliers(data, x, outlier.mode = "tukey", lower.percentile = 0.05,
  upper.percentile = 0.95, progress = TRUE)
}
\arguments{
\item{data}{[required | data.frame] Dataset containing features}

\item{x}{[required | character] A vector of numeric feature names present in the dataset}

\item{outlier.mode}{[optional | character | default="tukey"] Mode to identify outliers. Options are tukey or percentile to identify outliers.}

\item{lower.percentile}{[optional | numeric | default=0.05] Lower percentile used to identify outliers if mode is set to percentile}

\item{upper.percentile}{[optional | numeric | default=0.95] Upper percentile used to identify outliers if mode is set to percentile}

\item{progress}{[optional | logical | default=TRUE] Display a progress bar}
}
\value{
List of mapping tables
}
\description{
Identifies outliers by using either tukey or a percentile based approach. For each feature any value greater than or less than the outlier values are seen as outliers.
}
\examples{
om <- map.outliers(iris, x = names(iris)[1:4])
}
\author{
Xander Horn
}
