% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_interactions.R
\name{numeric.interactions}
\alias{numeric.interactions}
\title{Numeric feature interactions}
\usage{
numeric.interactions(data, x, progress = TRUE)
}
\arguments{
\item{data}{[required | data.frame] Dataset containing categorical features}

\item{x}{[required | character] A vector of numeric feature names present in the dataset}

\item{progress}{[optional | logical | default=TRUE] Display a progress bar}
}
\value{
Data frame or vector of interacted features
}
\description{
Computes two way feature interactions by summing, deducting, multiplying and dividing features.
}
\examples{
res <- numeric.interactions(iris, setdiff(names(iris), "Species"))
}
\author{
Xander Horn
}
