% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_transformers.R
\name{numeric.transformers}
\alias{numeric.transformers}
\title{Numeric feature transformations}
\usage{
numeric.transformers(data, x, transform.type = "log", progress = TRUE)
}
\arguments{
\item{data}{[required | data.frame] Dataset containing numeric features}

\item{x}{[required | character] A vector of numeric feature names present in the dataset}

\item{transform.type}{[optional | character | default="log"] Transform type, options are log or sqrt.}

\item{progress}{[optional | logical | default=TRUE] Display a progress bar}
}
\value{
Data frame containing transformed features
}
\description{
Applies different transformations to numeric features such as sqrt and log transforms.
}
\examples{
res <- numeric.transformers(data = iris, x = "Sepal.Length")
}
\author{
Xander Horns
}
