% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_format.R
\name{quick.format}
\alias{quick.format}
\title{Minor feature class changes}
\usage{
quick.format(data)
}
\arguments{
\item{data}{[required | data.frame] Dataset containing features}
}
\value{
Formatted dataset
}
\description{
Alters features in a dataset to change from factors to characters and logicals to numerics
}
\examples{
res <- quick.format(iris)
}
\author{
Xander Horn
}
