% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_features.R
\name{text.features}
\alias{text.features}
\title{Creates engineered features from text features}
\usage{
\method{text}{features}(data, x, progress = TRUE)
}
\arguments{
\item{data}{[required | data.frame] Dataset containing categorical features}

\item{x}{[required | character] A vector of text feature names present in the dataset}

\item{progress}{[optional | logical | default=TRUE] Display a progress bar}
}
\value{
Data frame with engineered text features
}
\description{
Simple features are created from text features such as punctuation count, character count, number count etc.
}
\examples{
tmp <- iris
tmp$Species <- as.character(tmp$Species)
tf <- text.features(data = tmp, x = "Species")
}
