% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_partition.R
\name{time.partition}
\alias{time.partition}
\title{Create time sensitive partitions}
\usage{
\method{time}{partition}(data, time.feature, valid.sample.size = 0.3)
}
\arguments{
\item{data}{[required | data.frame] Dataset to partition}

\item{time.feature}{[required | character] Time feature present in dataset}

\item{valid.sample.size}{[required | numeric | default=0.3]}
}
\value{
List with train and validation row indices
}
\description{
Partitions data according to a time based feature for time sensitive datasets
}
\examples{
new.iris <- iris
new.iris$date <- seq(ISOdate(2019,1,1), by = "day", length.out = nrow(new.iris))
splits <- time.partition(new.iris, "date")
}
\author{
Xander Horn
}
